https://colab.research.google.com/drive/15zaUu5t1js44A-KwwKJMGKS_N3m5RjHn?usp=sharing
Our project embarks on a pioneering journey into the realm of AI-powered music composition, where algorithms and machine learning converge to redefine the boundaries of musical creativity. In the ever-evolving landscape of technology, the marriage of artificial intelligence and music has emerged as a captivating frontier, pushing the boundaries of creativity and artistic expression. Our project, "Harmony in Code," delves into the realm of AI-powered music composition with a specific focus on the intricate and emotive world of piano music. In this we used the method called LSTM(Long Short-Term Memory)LSTMs are particularly well-suited for sequential data tasks such as time series prediction, natural language processing, and, relevant to your problem statement, music composition.

In the context of AI-powered music composition, LSTM networks can be used to learn patterns and structures within musical sequences and generate new compositions.